Create Convex hull
Convex Hull
easy
Time Limit: 2 sec
Memory Limit: 128000 kB
Problem Statement
Convex Hull of a set of points, in the 2D plane, is a convex polygon with the minimum area such that each point lies either on the boundary of the polygon or inside it. Now given a set of points the task is to find the convex hull of points.

Problem was asked by Ola, Samsung, Adobe, Goldman Sachs.
Input
Each test case contains an integer N denoting the no of points. Then in the next line are N*2 space-separated values denoting the points ie x and y.

Constraints:-
1 <= N <= 1000
1 <= x, y <= 1000
Output
For each test case in a new line print the points x and y of the convex hull separated by a space in sorted order where every pair is separated from the other by a ','. If no convex hull is possible to print -1.
Example
Sample Input:
3
1 2 3 1 5 6

Sample Output:
1 2, 3 1, 5 6

Sample Input
3
1 2 4 4 5 1

Sample Output:
1 2, 4 4, 5 1